Commit chat.js changes
Added support for session change via chat. This functionality is ideal for RPG modules which use a lot of different maps. In such a case, instead of making one huge module, each map can be made as a session and then GM can use this function to move all players to the desired session.